Added changes to allow delete on zero quantity
Deletes can now only be made on zero quantity. I don't know
if it makes sense to allow capabilities to ever force a delete
on an item without zero quantity, it might make sense for the
callee to modify the quantity to zero before deleting as a
statement of intent. Thoughts?